# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: CrossUniversityLibrary
  description: OpenAPI documentation for RESTful Service CrossUniversityLibrary
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Localhost Server URL
  - url: http://25.59.220.248:8080
    description: Hamachi Server URL
security:
  - bearerAuth: []
paths:
...
  /library/{libraryId}/unsubscribe:
    put:
      tags:
        - Library Controller
      summary: Unsubscribe from a library
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        unsubscribe from a library to which they are subscribed. Users can
        provide the library ID from which they want to unsubscribe. The response
        includes the updated list of libraries from which the user is
        unsubscribed, presented in the form of LibraryDTO (Data Transfer Object)
        instances.
      operationId: unsubscribeLibrary
      parameters:
        - name: libraryId
          in: path
          description: ID of the library from which the user wants to unsubscribe
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                libraries:
                  - id: 3
                    title: Student1_library1
                    topic: History
                    libraryAccess: false
                  - id: 5
                    title: Student2_library2
                    topic: Math
                    libraryAccess: false
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /library/{libraryId}/unsubscribe/{email}:
    put:
      tags:
        - Library Controller
      summary: Unsubscribe a user from a library
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        unsubscribe another user from a library they own. Users can provide the
        library ID and the email of the user they want to unsubscribe. The
        response is a standard 204 No Content status, indicating a successful
        operation without additional response data.
      operationId: unsubscribeUser
      parameters:
        - name: libraryId
          in: path
          description: >-
            ID of the library from which the user wants to unsubscribe another
            user
          required: true
          schema:
            type: integer
            format: int64
        - name: email
          in: path
          description: The email of the user to be unsubscribed
          required: true
          schema:
            type: string
      responses:
        '204':
          description: >-
            The request was successful, but there is no additional content to
            return. The server has processed the request, and no response body
            is provided.
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /library/{libraryId}/subscribe:
    put:
      tags:
        - Library Controller
      summary: Subscribe to a library
      description: >-
        This API endpoint is accessible to authenticated users, enabling them to
        subscribe to a library. Users can provide the library ID to which they
        want to subscribe. The response includes the updated list of libraries
        to which the user is subscribed, presented in the form of LibraryDTO
        (Data Transfer Object) instances.
      operationId: subscribeToLibrary
      parameters:
        - name: libraryId
          in: path
          description: ID of the library to which the user wants to subscribe
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                libraries:
                  - id: 3
                    title: Student1_library1
                    topic: History
                    libraryAccess: false
                  - id: 5
                    title: Student2_library2
                    topic: Math
                    libraryAccess: false
        '400':
          description: >-
            The server could not understand the request due to invalid or
            malformed parameters.
          content:
            application/json:
              example:
                message: Bad request example message
                httpStatus: BAD_REQUEST
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /library/{libraryId}/subscribe/{email}:
    put:
      tags:
        - Library Controller
      summary: Subscribe a user to a library
      description: >-
        This API endpoint is accessible to authenticated users, enabling them to
        subscribe another user to a library they own. Users can provide the
        library ID and the email of the user they want to subscribe. The
        response is a standard 204 No Content status, indicating a successful
        operation without additional response data.
      operationId: subscribeUser
      parameters:
        - name: libraryId
          in: path
          description: ID of the library to which the user wants to subscribe another user
          required: true
          schema:
            type: integer
            format: int64
        - name: email
          in: path
          description: The email of the user to be subscribed
          required: true
          schema:
            type: string
      responses:
        '204':
          description: >-
            The request was successful, but there is no additional content to
            return. The server has processed the request, and no response body
            is provided.
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: >-
            The server could not understand the request due to invalid or
            malformed parameters.
          content:
            application/json:
              example:
                message: Bad request example message
                httpStatus: BAD_REQUEST
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /library/{libraryId}/document/{documentId}/remove:
    put:
      tags:
        - Library Controller
      summary: Remove an existing document from a library
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        remove an existing document from their library, specifically documents
        that were added from other libraries. Users can provide the library ID
        and the document ID to be removed. The response is a standard 204 No
        Content status, indicating a successful operation without additional
        response data.
      operationId: removeDocumentFromLibrary
      parameters:
        - name: libraryId
          in: path
          description: >-
            ID of the library from which the user wants to remove an existing
            document
          required: true
          schema:
            type: integer
            format: int64
        - name: documentId
          in: path
          description: ID of the existing document to be removed
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: >-
            The request was successful, but there is no additional content to
            return. The server has processed the request, and no response body
            is provided.
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /library/{libraryId}/document/{documentId}/add:
    put:
      tags:
        - Library Controller
      summary: Add an existing document to a library
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        add an existing document from other libraries to libraries they own.
        Users can provide the library ID and the document ID to be added. The
        response is a standard 204 No Content status, indicating a successful
        operation without additional response data.
      operationId: addExistedDocumentToLibrary
      parameters:
        - name: libraryId
          in: path
          description: >-
            ID of the library to which the user wants to add an existing
            document
          required: true
          schema:
            type: integer
            format: int64
        - name: documentId
          in: path
          description: ID of the existing document to be added
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: >-
            The request was successful, but there is no additional content to
            return. The server has processed the request, and no response body
            is provided.
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /library/update:
    put:
      tags:
        - Library Controller
      summary: Update library information
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        update the title and topic of a library they own. Users can provide a
        LibraryDTO (Data Transfer Object) containing the updated information.
        The response includes the updated LibraryDTO instance representing the
        modified library.
      operationId: updateLibrary
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryDTO'
            example:
              id: 3
              title: library1_title
              topic: History
              libraryAccess: false
        required: true
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                id: 3
                title: library1_title
                topic: History
                libraryAccess: false
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /library/document/update:
    put:
      tags:
        - Library Controller
      summary: Update document information
      description: >-
        This API endpoint is accessible to authenticated users, specifically
        allowing the owner to update information about a document, including the
        title, topic, and description. Users can provide a DocumentDTO (Data
        Transfer Object) containing the updated information. The response
        includes the updated DocumentDTO instance representing the modified
        document.
      operationId: updateDocument
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentDTO'
            example:
              id: 4
              title: Document 4
              topic: Math Science
              description: Description for Document 4
        required: true
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                id: 4
                title: Document 4
                topic: Math Science
                description: Description for Document 4
                owner:
                  id: 1
                  userName: student1_name
                  email: student1kpi@kpi.ua
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
...
  /library/create:
    post:
      tags:
        - Library Controller
      summary: Create a library with default closed access
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        create a new library specifically for students. Libraries created by
        students always have restricted access. The response includes relevant
        details about the created library, presented in the form of LibraryDTO
        (Data Transfer Object) instances
      operationId: createLibrary
      parameters:
        - name: title
          in: query
          description: The title of the library
          required: true
          schema:
            type: string
        - name: topic
          in: query
          description: The topic of the library
          required: true
          schema:
            type: string
      responses:
        '201':
          description: >-
            The resource has been successfully created. The server has processed
            the request to create a new resource, and the operation was
            completed without errors.
          content:
            application/json:
              example:
                id: 3
                title: library1_title
                topic: History
                libraryAccess: false
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
...
components:
  schemas:
    LibraryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3
        title:
          type: string
          example: library_title
        topic:
          type: string
          example: library_topic
        libraryAccess:
          type: boolean
          example: false
      description: An object representing a library
    DocumentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        topic:
          type: string
        description:
          type: string
        owner:
          $ref: '#/components/schemas/UserBriefProfile'
      description: An object representing a document
    UserBriefProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userName:
          type: string
        email:
          type: string
      description: An object representing a brief user profile
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT authentication
      scheme: bearer
      bearerFormat: JWT